First attempt at working matcher
